tungnv_b23dckh130_training_gd1/Week 3/BE/Practice/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt
